+++ frontend/src/helper.js
@@ -43,7 +43,6 @@ const getCurrencyFormat = currency => ({
   ...currencies[currency]
 });
 
-// TODO: come up with a better solution for the amount not requiring to parse every time
 export const compareObjects = (items, itemToAdd) => {
   if (!_isEmpty(items)) {
     const itemToAddClone = _cloneDeep(itemToAdd);
@@ -78,11 +77,9 @@ export const formatAmountString = (amount, currency) => {
   }
   return amount;
 };
-export const getCurrencies = parentCurrency => {
+export const getCurrencies = () => {
   return Object.keys(currencies).map(currency => {
-    const disabled = !_isEmpty(parentCurrency) && !(parentCurrency === currency);
     return {
-      disabled,
       primaryText: currency,
       value: currency
     };
@@ -145,7 +142,7 @@ export const roleMapper = {
   assignee: strings.common.assignee
 };
 
-const createDoughnutData = (labels, data, colors = statusColors) => ({
+export const createDoughnutData = (labels, data, colors = statusColors) => ({
   labels,
   datasets: [
     {
@@ -197,10 +194,6 @@ export const formatUpdateString = (identifier, createdBy, data) => {
   return string.concat(changes);
 };
 
-export const getAllocationRatio = (spentAmount, projectAmount) => {
-  const allocationRatio = spentAmount / projectAmount * 100;
-  return allocationRatio > 0 ? allocationRatio : 0;
-};
 export const calculateWorkflowBudget = workflows => {
   return workflows.reduce(
     (acc, workflow) => {