+++ frontend/src/pages/Overview/ProjectDialog.js
@@ -45,6 +45,9 @@ const ProjectDialog = props => {
   const { projects, projectToAdd, editDialogShown, creationDialogShown } = props;
   const { displayName } = projectToAdd;
   const changes = compareObjects(projects, projectToAdd);
+  if (changes.deletedProjectedBudgets && changes.deletedProjectedBudgets.length === 0) {
+    delete changes.deletedProjectedBudgets;
+  }
   const specificProps = props.editDialogShown
     ? {
         handleSubmit: handleEdit,
@@ -59,7 +62,7 @@ const ProjectDialog = props => {
     {
       title: strings.project.project_details,
       content: <ProjectDialogContent {...props} />,
-      nextDisabled: _isEmpty(displayName) || _isEmpty(changes)
+      nextDisabled: _isEmpty(displayName) || (_isEmpty(changes) && props.editDialogShown)
     }
   ];