+++ frontend/src/pages/SubProjects/SubprojectDialogContent.js
@@ -20,7 +20,7 @@ function getMenuItems(currencies) {
 }
 
 const SubprojectDialogContent = props => {
-  const currencies = getCurrencies(props.projectCurrency);
+  const currencies = getCurrencies();
   return (
     <div>
       <div>
@@ -50,7 +50,6 @@ const SubprojectDialogContent = props => {
       <div>
         <Budget
           currencyTitle={strings.subproject.subproject_currency}
-          parentCurrency={props.projectCurrency}
           budgetLabel={strings.subproject.subproject_budget_amount}
           projectedBudgets={props.subprojectToAdd.projectedBudgets}
           storeProjectedBudget={props.storeSubProjectProjectedBudgets}