+++ frontend/src/sagas.js
@@ -145,6 +145,18 @@ import {
   EXPORT_DATA_SUCCESS,
   EXPORT_DATA_FAILED
 } from "./pages/Navbar/actions.js";
+import {
+  GET_SUBPROJECT_KPIS,
+  GET_SUBPROJECT_KPIS_SUCCESS,
+  GET_PROJECT_KPIS,
+  GET_PROJECT_KPIS_SUCCESS,
+  GET_EXCHANGE_RATES,
+  GET_EXCHANGE_RATES_SUCCESS,
+  GET_SUBPROJECT_KPIS_FAIL,
+  GET_PROJECT_KPIS_FAIL
+} from "./pages/Analytics/actions.js";
+import { fromAmountString } from "./helper.js";
+import { getExchangeRates } from "./getExchangeRates";
 
 const api = new Api();
 
@@ -1224,6 +1236,181 @@ export function* liveUpdateNotificationsSaga({ showLoading, offset }) {
   }, showLoading);
 }
 
+export function* getProjectKPIsSaga({ projectId, showLoading = true }) {
+  yield execute(function*() {
+    const {
+      data: {
+        project: {
+          data: { projectedBudgets }
+        },
+        subprojects
+      }
+    } = yield callApi(api.viewProjectDetails, projectId);
+
+    try {
+      const subprojectBudgets = (yield all(
+        subprojects.map(subproject => callApi(api.viewSubProjectDetails, projectId, subproject.data.id))
+      )).map(subprojectDetails => {
+        const currency = subprojectDetails.data.subproject.data.currency;
+        const projected = subprojectDetails.data.subproject.data.projectedBudgets;
+        const workflowBudgets = subprojectDetails.data.workflowitems.reduce(
+          (acc, next) => {
+            if (!next.data.amountType) {
+              const error = new Error("redacted");
+              error.name = "redacted";
+              throw error;
+            }
+            const { amountType, status, amount, exchangeRate } = next.data;
+            if (amountType === "allocated" && amount) {
+              return {
+                ...acc,
+                allocated: acc.allocated + fromAmountString(amount) * (exchangeRate || 1)
+              };
+            }
+
+            if (amountType === "disbursed" && amount) {
+              return {
+                ...acc,
+                disbursed:
+                  status === "closed" ? acc.disbursed + fromAmountString(amount) * (exchangeRate || 1) : acc.disbursed
+              };
+            }
+
+            return acc;
+          },
+          {
+            allocated: 0,
+            disbursed: 0
+          }
+        );
+        return {
+          projected,
+          currency,
+          disbursed: workflowBudgets.disbursed,
+          allocated: workflowBudgets.allocated
+        };
+      });
+      const projectBudgets = subprojectBudgets.reduce(
+        (acc, next) => {
+          if (next.disbursed !== 0) {
+            acc.disbursed.push({ budget: next.disbursed, currency: next.currency });
+          }
+          if (next.allocated !== 0) {
+            acc.allocated.push({ budget: next.allocated, currency: next.currency });
+          }
+          if (next.projected.length !== 0) {
+            acc.projectedOfSubprojects.push(next.projected);
+          }
+          return {
+            disbursed: acc.disbursed,
+            allocated: acc.allocated,
+            projectedOfSubprojects: acc.projectedOfSubprojects
+          };
+        },
+        { disbursed: [], allocated: [], projectedOfSubprojects: [] }
+      );
+
+      yield put({
+        type: GET_EXCHANGE_RATES,
+        baseCurrency: projectedBudgets[0] ? projectedBudgets[0].currencyCode : "EUR"
+      });
+
+      yield put({
+        type: GET_PROJECT_KPIS_SUCCESS,
+        assignedBudget: projectBudgets.allocated,
+        disbursedBudget: projectBudgets.disbursed,
+        projectedBudget: projectBudgets.projectedOfSubprojects,
+        totalBudget: projectedBudgets,
+        displayCurrency: projectedBudgets[0] ? projectedBudgets[0].currencyCode : "EUR"
+      });
+    } catch (error) {
+      if (error.name === "redacted") {
+        yield put({
+          type: GET_PROJECT_KPIS_FAIL,
+          reason: "redacted"
+        });
+      } else {
+        throw error;
+      }
+    }
+  }, showLoading);
+}
+
+export function* getSubProjectKPIs({ projectId, subProjectId, showLoading = true }) {
+  yield execute(function*() {
+    const {
+      data: {
+        workflowitems = [],
+        subproject: {
+          data: { projectedBudgets = [], currency = "EUR" }
+        }
+      }
+    } = yield callApi(api.viewSubProjectDetails, projectId, subProjectId);
+    yield put({
+      type: GET_EXCHANGE_RATES,
+      baseCurrency: currency
+    });
+    try {
+      const workflowBudgets = workflowitems.reduce(
+        (acc, next) => {
+          if (!next.data.amountType) {
+            const error = new Error("redacted");
+            error.name = "redacted";
+            throw error;
+          }
+          const { amountType, status, amount, exchangeRate } = next.data;
+          if (amountType === "allocated" && status === "closed" && amount) {
+            return {
+              ...acc,
+              assignedBudget: acc.assignedBudget + fromAmountString(amount) * exchangeRate
+            };
+          }
+
+          if (amountType === "disbursed" && status === "closed" && amount) {
+            return {
+              ...acc,
+              disbursedBudget: acc.disbursedBudget + fromAmountString(amount) * exchangeRate
+            };
+          }
+
+          return acc;
+        },
+        { assignedBudget: 0, disbursedBudget: 0 }
+      );
+
+      const response = {
+        subProjectCurrency: currency,
+        projectedBudgets: projectedBudgets,
+        assignedBudget: workflowBudgets.assignedBudget,
+        disbursedBudget: workflowBudgets.disbursedBudget
+      };
+      yield put({
+        type: GET_SUBPROJECT_KPIS_SUCCESS,
+        ...response
+      });
+    } catch (error) {
+      if (error.name === "redacted") {
+        yield put({
+          type: GET_SUBPROJECT_KPIS_FAIL,
+          reason: "redacted"
+        });
+      } else {
+        throw error;
+      }
+    }
+  }, showLoading);
+}
+
+export function* getExchangeRatesSaga({ baseCurrency, showLoading = true }) {
+  yield execute(function*() {
+    const exchangeRates = yield getExchangeRates(baseCurrency);
+    yield put({
+      type: GET_EXCHANGE_RATES_SUCCESS,
+      exchangeRates
+    });
+  }, showLoading);
+}
+
 function* exportDataSaga() {
   yield execute(
     function*() {
@@ -1329,6 +1516,11 @@ export default function* rootSaga() {
       yield takeLatest(CREATE_BACKUP, createBackupSaga),
       yield takeLatest(RESTORE_BACKUP, restoreBackupSaga),
       yield takeLatest(FETCH_VERSIONS, fetchVersionsSaga),
+
+      // Analytics
+      yield takeLeading(GET_SUBPROJECT_KPIS, getSubProjectKPIs),
+      yield takeLeading(GET_PROJECT_KPIS, getProjectKPIsSaga),
+      yield takeLeading(GET_EXCHANGE_RATES, getExchangeRatesSaga),
       yield takeLeading(EXPORT_DATA, exportDataSaga)
     ]);
   } catch (error) {